+++ clients/js/jest.config.mjs
@@ -3,6 +3,19 @@ const config = {
   coverageDirectory: 'coverage/jest',
   coveragePathIgnorePatterns: ['test/*'],
   coverageReporters: ['lcov', 'text', 'cobertura', 'html'],
+  /*
+  coverageThreshold: {
+    global: {
+      branches: 80,
+      functions: 80,
+      lines: 80,
+      statements: 80,
+    },
+    'src/networks.ts': {
+      branches: 50
+    }
+  },
+  */
   moduleFileExtensions: ['js', 'ts'],
   moduleNameMapper: {
     '^@oasisprotocol/sapphire-paratime$': '<rootDir>/src/index',